asm: remove header from binary files.
Binary files are now always raw data. All metadata is moved to the
source map file.  This means the source map file now includes the
origin address and size of the corresponding binary file.  I also
added a CRC of the binary file so the map can be marked as invalid
if the binary changes without updating the map file.

Added a new host command: 'assemble map'. This allows you to create
an empty source map file for an existing binary so it's easier to
load.